Change the tests to Property Tests against the .NET implementation * Improve precision on FloatAverageAggregator

+++ Funcky/Internal/Aggregators/FloatAverageAggregator.cs
@@ -6,13 +6,15 @@ namespace Funcky.Internal.Aggregators
     {
         public static readonly FloatAverageAggregator Empty = new();
 
-        private readonly Option<float> _sum;
+        // we calculate with double for higher precision and fewer overflow problems
+        // The implementation in .NET seems to do the same, because we seem to have no aberration.
+        private readonly Option<double> _sum;
         private readonly int _count;
 
-        private FloatAverageAggregator(int count = default, Option<float> sum = default)
+        private FloatAverageAggregator(int count = default, Option<double> sum = default)
             => (_count, _sum) = (count, sum);
 
-        public Option<float> Average => _sum.Select(sum => sum / _count);
+        public Option<float> Average => _sum.Select(sum => (float)(sum / _count));
 
         public FloatAverageAggregator Add(float term)
             => new(_count + 1, Some(_sum.Match(none: term, some: sum => sum + term)));
